Port current API proposal to Smithy
This commit models the current Autopeer API using Smithy as
the Interface Description Language prior to converting it to
Openapi 3.0.2.

This model follows a REST design with 2 resources:
- location describes a physical location where peering takes place
- session describes the logical connection between entities prior
  to traffic exchange

The API uses common representations for a CRUDL REST API although only
creation, retrieval and listing are currently proposed.

It also uses well-known good practices to model the REST interface
such as idempotency markers on a server-generated resource and
pagination of possibly long list of resources.

This commit also introduces usage examples for all API resources and
verbs in operation.